backend/requirements.txt
flask
pymongo
python-dotenv
requests
flask-cors
gunicorn
groq
pdfplumber